application-tier/models/model-ocr/ocr_requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
aiohttp==3.9.1
httpx==0.25.2
requests==2.31.0
azure-ai-formrecognizer==3.3.0
azure-core==1.29.5
